Extract new data types to a separate package
